moved ContentNode class to own module
