+++ tests/node.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 class ContentNodeTests(TestCase):
 
     def test_follows_buildPage(self):
-        from syrinx.read import ContentNode
+        from syrinx.node import ContentNode
         node = ContentNode(None, None)
         node.name = 'foo_bar'
         node.front = dict()
